Move functions into lib so they can be reused
